完善日志信息 修改b站获取开播状态的接口 v0.2.1

+++ cmd/bot/config.go
@@ -7,7 +7,7 @@ import (
 )
 
 type BiliCfg struct {
-	Live    []int64 `yaml:"live"`
+	Live    []int   `yaml:"live"`
 	Dynamic []int64 `yaml:"dynamic"`
 }
